Move Fenced into super
